Moved display code to seperate file
